Use generic GSS functions, not krb5 API
Let us no longer use the krb5 API for GSS things.  Instead we now use
the newer gss_store_cred() for storing delegated credentials (to the
default credential store), and gss_userok() and gss_localname() for
authorization tasks.

Now, we lose the ability to have per-session credential stores, and
along with that, the ability to destroy them on logout.  This is fine
because at least for our users, the pain of having to deal with
different credential stores on screen/tmux re-attach is too great to
tolerate.

However, if anyone needs to regain per-session credential stores,
well, there's no GSS equivalent of krb5_cc_new_unique().  There is
gss_store_cred_into(), but it cannot generate a unique credential
store, so the caller would have to, and that would a mechanism-
specific thing ("ccname","<mkstempo()'ed-name>").  What we really
want is a new function we might call gss_store_cred_into_new() that
takes a const store key/value set, outputs a new one, and maybe also
outputs a set of envvars to set, and then we'd need release
functions, gss_release_store() and gss_release_env(), for those
things, and lastly, a gss_store_destroy() for logout.