Fix GSS KEYEX hostkey handling (buf. is read-only)
We can't use sshpkt_getb_froms() because that retains a reference to the
buffer containing the current packet.  Otherwise connecting to a server
that offers a server host public key causes the client to die with:

    Received KEXGSS_HOSTKEY
    internal error: buffer is read-only

Ideally there would be a corresponding function that gets a copy, or
else reading the next packet should allocate a new internal buffer and
cause the one returned by sshpkt_getb_froms() to eventually free the
original internal buffer.